// web/src/lib/schemas.ts
export const AccountBase = z.object({
  name: z.string(),
  type: z.enum(["savings", "checking", "creditCard"]),
  balance: z.number(),
  userId: IdSchema,
});

export const AccountSchema = addIdAndDatetime(AccountBase);
export type Account = z.infer<typeof AccountSchema>;

// Budget
export const BudgetBase = z.object({
  month: z.number(),
  year: z.number(),
  userId: IdSchema,
});

export const BudgetSchema = addIdAndDatetime(BudgetBase);
export type Budget = z.infer<typeof BudgetSchema>;

// Budget Item
export const BudgetItemBase = z.object({
  actualAmount: z.number(),
  targetAmount: z.number(),
  budgetId: IdSchema,
  categoryId: IdSchema,
});

export const BudgetItemSchema = addIdAndDatetime(BudgetItemBase);
export type BudgetItem = z.infer<typeof BudgetItemSchema>;

// Category
export const CategoryBase = z.object({
  name: z.string(),
  group: z.string(),
  status: z.string(),
  userId: IdSchema,
});

export const CategorySchema = addIdAndDatetime(CategoryBase);
export type Category = z.infer<typeof CategorySchema>;

// Transaction
export const TransactionBase = z.object({
  merchant: z.string(),
  amount: z.number(),
  notes: z.string(),
  date: z.coerce.date(),
  userId: IdSchema,
  accountId: IdSchema,
  budgetItemId: IdSchema,
  type: z.enum(["credit", "debit"]),
  recurringTemplateId: z.nullish(IdSchema),
});

export const TransactionSchema = addIdAndDatetime(TransactionBase);
export type Transaction = z.infer<typeof TransactionSchema>;